BeginingMVC4
What Is ASP.NET?

ASP.NET is a Web platform that provides all the services that you
require to build enterprise-class server-based Web applications. ASP.NET
is built on the .NET Framework, so all .NET Framework features are
available to ASP.NET applications. Your applications can be written in
any language that is compatible with the common language runtime (CLR),
including Visual Basic and C#. To create ASP.NET Web applications, you
can use Visual Studio. The tools and options in Visual Studio that are
designed for creating Web applications are referred to collectively as
Visual Web Developer. In addition, a free standalone product—Visual Web
Developer Express—is available that includes the core set of Web-design
features from Visual Studio.

ASP.NET includes the following features:
- A page and controls framework
- The ASP.NET compiler
- Security infrastructure
- State-management facilities
- Application configuration
- Health monitoring and performance features
- Debugging support
- An XML web services framework, which was later superseded by Windows
Communication
- Foundation (WCF) An extensible hosting environment and application
lifecycle management
- An extensible designer environment

ASP.NET Web Forms

ASP.NET Web Forms allows you to create web applications in the same way
you would create a traditional Windows Forms application. This means
that you have a control-based interface and the following two files for
each page:
• A user interface (UI) file: Includes the markup and is where you
design how your page will look and which controls it will use. This file
has the extension .aspx.
• A code-behind file: Includes the code that handles events and
interactions of all the objects in the page (this code could be included
on the preceding .aspx page, but normally is in a separate file). This
file has an extension associated with the programming language, either
aspx.cs for C# or aspx.vb for VB.NET.

When to Use ASP.NET MVC

ASP.NET MVC has certain capabilities that make it the best option to
choose if you need one or more of the following:
• A high level of control over the generated HTML: Unlike Web Forms,
Views in ASP.NET MVC render HTML exactly as you tell them to. Recently,
Web Forms have been improved in this area but still don’t have the level
of control MVC has.
• Easier unit testing: With ASP.NET MVC, it is very easy to follow
testing patterns such as  test-driven development (TDD). Because of the
complex event lifecycle in Web Forms,  on top of a control-based
framework, TDD is a lot easier with MVC.
• Separation of concerns: This refers to having all aspects of the
system clearly separated from one another. Because of the pattern it
implements, an MVC application is divided into discrete and loosely
bound parts (model, views, and controllers), which makes it easy to
maintain

ASP.NET MVC Benefits
Compared to Web Forms, ASP.NET MVC applications benefit by including all
ASP.NET core features but also by the features in the MVC pattern. Some
of those benefits are:
- The MVC pattern itself makes it easier to manage complexity by clearly
separating the  functionality of the application into three core parts,
the model, the view, and the controller.
- ASP.NET MVC web applications do not use view state or server-based
forms. This makes the • MVC framework ideal for developers who want full
control over the behavior of an application. View state can become very
large, which is a problem for devices like smartphones running over slow
networks (transmitting all that information can be very slow). In a Web
Forms page, you could only have one <form> per page. This is quite a
major restriction. In MVC, there is no such restriction—that is, you can
have as many<form> elements as you like.
- ASP.NET MVC provides better support for test-driven development (TDD).
- ASP.NET MVC works well for web applications that are supported by
large teams of developers and for web designers who need a high degree
of control over the HTML.

ASP.NET MVC 4 Features ASP.NET MVC 4 is built on top of the many
features of its previous versions and include new features like:
- ASP.NET Web API, a new framework for building HTTP and RESTful
services.
- A new HTML5-based default template in Visual Studio and a new Mobile
Application project template.
- Automatic selection of rendered views with Display Modes. This is
particularly useful when  building applications that will run not only
on desktop browsers but on mobile browsers as well. It will let the
application determine the best view to render based on the browser
making the request.
- jQuery Mobile and mobile features.
- Task support for asynchronous controllers.
- Microsoft Windows Azure SDK support for deploying ASP.NET MVC 4
applications to  Windows Azure.
- Bundling and minification for CSS and JavaScript files to help reduce
the amount of HTTP requests as well as the size of those requests.
- Facebook, OpenID, and OAuth authentication.

Summary

You have seen in this introductory chapter what ASP.NET is and the core
technologies it contains: Web Forms,  Web Pages, and MVC. While Web
Forms and Web Pages are simply other options to create web applications,
MVC has some advantages and strengths that make it a better fit for
certain scenarios, such as where you need more control over the
generated HTML and better support for implementing unit testing. You
have seen that the way in which MVC handles requests differs from how
Web Forms and Web Pages handle requests. This is a very important
concept, as it highlights the decoupling of the request from the
physical page,  using a powerful routing engine. Among the features in
ASP.NET MVC 4 is Web API, a powerful new framework for building HTTP and
RESTful services. A mobile template is included in ASP.NET MVC 4 so that
you can support mobile browsers out of the box. Asynchronous controllers
can be implemented using tasks. Easy Facebook and OAuth implementations
are also included to create a friendlier authentication scheme.
Integration with Windows Azure via the Windows Azure SDK is now
available, as well as bundling and minification of CSS and JavaScript
files to improve speed and bandwidth consumption. Having read this
chapter, you now have the information to know when to choose ASP.NET MVC
as the technology for your project. ASP.NET MVC is a better fit if you
want a clear separation of concerns by having a model, controllers and
views thus reducing the complexity of the project. Also, ASP.NET MVC
works well if you have a large team with people doing specific tasks.